requirements.txt
pillow==11.2.1
py-tgcalls==2.1.1
pymongo==4.12.1
python-dotenv==1.1.0
Telethon==1.40.0
yt-dlp==2025.3.31